Mike pointed out these should be u32.
[!no_src_qa!] - extern in C file is fine.



git-svn-id: svn://localhost/svn/linux-kernel/trunk@6571 526b6c2d-f592-4532-a319-5dd88ccb003d

+++ arch/blackfin/kernel/traps.c
@@ -70,7 +70,7 @@
 #endif
 
 #ifdef CONFIG_DEBUG_MMRS
-int last_seqstat;
+u32 last_seqstat;
 #endif
 
 /* Initiate the event table handler */
@@ -251,7 +251,7 @@ asmlinkage void trap_c(struct pt_regs *fp)
 
 	trace_buffer_save(j);
 #ifdef CONFIG_DEBUG_MMRS
-	last_seqstat = fp->seqstat;
+	last_seqstat = (u32)fp->seqstat;
 #endif
 
 	/* Important - be very careful dereferncing pointers - will lead to
